+++ testapi/views.py
@@ -13,7 +13,10 @@ from django.conf import settings
 from django.core.signing import Signer
 from django.db.models import Q
 
+from buyer.models import Buyer
+from buyer.serializers import BuyerSerializer
 from company.models import Company
+from core.authentication import Oauth2AuthenticationSSO
 from testapi.serializers import (
     CompanySerializer,
     ISDCompanySerializer,
@@ -28,6 +31,15 @@ class TestAPIView(GenericAPIView):
         if not settings.FEATURE_TEST_API_ENABLED:
             raise Http404
         return super().dispatch(*args, **kwargs)
+
+
+class BuyerTestAPIView(TestAPIView, RetrieveAPIView):
+    serializer_class = BuyerSerializer
+    authentication_classes = [Oauth2AuthenticationSSO]
+    permission_classes = []
+    queryset = Buyer.objects.all()
+    lookup_field = 'email'
+    http_method_names = 'get'
 
 
 class CompanyTestAPIView(TestAPIView, RetrieveAPIView, DestroyAPIView, UpdateAPIView):
@@ -119,5 +131,5 @@ class UnpublishedCompaniesTestAPIView(TestAPIView, RetrieveAPIView):
 
 class ISDCompanyTestAPIView(TestAPIView, CreateAPIView):
     serializer_class = ISDCompanySerializer
-    authentication_classes = []
+    authentication_classes = [Oauth2AuthenticationSSO]
     permission_classes = []